requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
flasgger==0.9.7.1
